fix: ensure visibility purge job correctly stops/restarts on Tomcat restart
Split the OSGi registration out of VisibilityActionPurgeJob into a
dedicated VisibilityActionPurgeJobRegistration component. The job is now
registered with a fixed name so it can be retrieved after a restart, the
stale job is deleted at startup, and unscheduling is skipped when the
scheduler is already stopped during shutdown.

Relates to Jahia/jahia-private#4978
